refactor: common test 분리
