+++ common/src/test/java/kitchenpos/common/domain/MoneyTest.java
@@ -1,7 +1,7 @@
-package kitchenpos.domain.common;
+package kitchenpos.common.domain;
 
-import kitchenpos.common.domain.Money;
-import kitchenpos.domain.DomainTest;
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -11,7 +11,8 @@ import java.math.BigDecimal;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class MoneyTest extends DomainTest {
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
+class MoneyTest {
     @Test
     void equals() {
         // given
