+++ common/src/test/java/kitchenpos/common/domain/QuantityTest.java
@@ -1,12 +1,13 @@
-package kitchenpos.domain.common;
+package kitchenpos.common.domain;
 
-import kitchenpos.common.domain.Quantity;
-import kitchenpos.domain.DomainTest;
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class QuantityTest extends DomainTest {
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
+class QuantityTest {
     @Test
     void throw_when_quantity_is_below_zero() {
         // when & then
